+++ 우석/jpashop/src/main/java/jpabook/jpashop/service/MemberService.java
@@ -44,12 +44,12 @@ public class MemberService {
 
     // 회원 단건 조회
     public Member findOne(Long memberId){
-        return memberRepository.findOne(memberId);
+        return memberRepository.findById(memberId).get();
     }
 
     @Transactional
     public void update(Long id, String name) {
-        Member member = memberRepository.findOne(id);
+        Member member = memberRepository.findById(id).get();
         member.setName(name);
     }
 }
